requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
boto3>=1.34.0
python-dotenv>=1.0.0
openai>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
httpx>=0.25.0
python-multipart>=0.0.6
PyPDF2>=3.0.1